maj roles admin

+++ controllers/comment.js
@@ -1,4 +1,5 @@
 const commentService = require('../services/comment');
+const h = require('./helpers');
 
 exports.findAll = async function (req, res) {
     console.log("comment.findAll");
@@ -6,7 +7,7 @@ exports.findAll = async function (req, res) {
         const comments = await commentService.findAll()
         return res.status(200).json(comments)
     } catch (error) {
-        return res.status(400).json(error.message)
+        return res.status(400).json({error : error.message})
     }
 };
 
@@ -17,38 +18,36 @@ exports.find = async function (req, res) {
         const comment = await commentService.find(id)
         return res.status(200).json(comment)
     } catch (error) {
-        return res.status(400).json(error.message)
+        return res.status(400).json({error:error.message})
     }
 };
 
 exports.create = async function (req, res) {
     console.log("comment.create");
     let pComment = req.body;
-    pComment.userId = res.locals.user;
-    // pComment.postId = parseInt(req.body.postId);
-    console.log("ctrl comment: ", pComment)
-
-    console.log(pComment);
+    pComment.userId = res.locals.user.id;
 
     try {
         const comment = await commentService.create(pComment)
         return res.status(200).json(comment)
     } catch (error) {
-        return res.status(400).json(error.message)
+        return res.status(400).json({error:error.message})
     }
 };
 
 exports.update = async function (req, res) {
     console.log("comment.update");
     const id = req.params.id
-    const comment = {
-        email : req.body.email
-    }
+    const message = req.body.message;
+    
     try {
-        const comments = await commentService.update(comment,id)
-        return res.status(200).json(comments)
+        const comment = await commentService.findByPk(id)
+        if( ! h.isOwnerOrAdmin(comment.userId, res.locals.user)) throw new Error("not your comment")
+
+        const result = await commentService.update(message,id)
+        return res.status(200).json(result)
     } catch (error) {
-        return res.status(400).json(error.message)
+        return res.status(400).json({error:error.message})
     }
 };
 
@@ -57,9 +56,13 @@ exports.delete = async function (req, res) {
     const id = req.params.id;
 
     try {
-        const comments = await commentService.delete(id)
-        return res.status(200).json(comments)
+
+        const comment = await commentService.findByPk(id)
+        if( ! h.isOwnerOrAdmin(comment.userId, res.locals.user)) throw new Error("not your comment")
+
+        const result = await commentService.delete(id)
+        return res.status(200).json(result)
     } catch (error) {
-        return res.status(400).json(error.message)
+        return res.status(400).json({error:error.message})
     }
 };
