+++ controllers/post.js
@@ -1,5 +1,7 @@
 const postService = require('../services/post');
 const userService = require('../services/user');
+const likeService = require('../services/like');
+const h = require('./helpers');
 
 exports.findAll = async function (req, res) {
     console.log("post.findAll");
@@ -13,7 +15,9 @@ exports.findAll = async function (req, res) {
 
 exports.find = async function (req, res) {
     console.log("post.find");
-    const id = {id:req.params.id};
+    // console.log(req.path);
+    // console.log(req.originalUrl) 
+    const id = { id: req.params.id };
     try {
         const post = await postService.find(id);
         return res.status(200).json(post)
@@ -25,8 +29,7 @@ exports.find = async function (req, res) {
 exports.create = async function (req, res) {
     console.log("post.create");
     const pPost = req.body;
-    // console.log("controler create req", res.locals.user);
-    pPost.userId = res.locals.user;
+    pPost.userId = res.locals.user.id;
     try {
         const post = await postService.create(pPost)
         return res.status(200).json(post)
@@ -35,17 +38,50 @@ exports.create = async function (req, res) {
     }
 };
 
+exports.like = async function (req, res) {
+    console.log("post.like");
+    const pPost = { postId: req.params.id };
+    pPost.userId = res.locals.user.id;
+    try {
+        const like = await likeService.create(pPost)
+        return res.status(200).json(like)
+    } catch (error) {
+        return res.status(400).json(error.message)
+    }
+};
+
+exports.unlike = async function (req, res) {
+    console.log("post.unlike");
+    const pPost = { postId: req.params.id };
+    pPost.userId = res.locals.user.id;
+    
+    try {
+        const like = await likeService.find(pPost)
+        if( ! h.isOwnerOrAdmin(like.userId, res.locals.user)) throw new Error("not your like")
+
+        const result = await likeService.delete(pPost)
+        return res.status(200).json(result)
+    } catch (error) {
+        return res.status(400).json(error.message)
+    }
+};
+
 exports.update = async function (req, res) {
     console.log("post.update");
     const id = req.params.id
-    const post = {
-        email : req.body.email
+    const body = {
+        message: req.body.message,
+        titre: req.body.titre
     }
     try {
-        const posts = await postService.update(post,id)
-        return res.status(200).json(posts)
+
+        const post = await postService.find({id:id})
+        if( ! h.isOwnerOrAdmin(post.userId, res.locals.user)) throw new Error("not your post") 
+
+        const result = await postService.update(body, id)
+        return res.status(200).json(result)
     } catch (error) {
-        return res.status(400).json(error.message)
+        return res.status(400).json({error: error.message})
     }
 };
 
@@ -54,9 +90,12 @@ exports.delete = async function (req, res) {
     const id = req.params.id;
 
     try {
-        const posts = await postService.delete(id)
-        return res.status(200).json(posts)
+        const post = await postService.find({id:id})
+        if( ! h.isOwnerOrAdmin(post.userId, res.locals.user)) throw new Error("not your post") 
+
+        const result = await postService.delete(id)
+        return res.status(200).json(result)
     } catch (error) {
-        return res.status(400).json(error.message)
+        return res.status(400).json({error: error.message})
     }
 };
